% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkingconditions.R
\name{check_fun_groupsparsw}
\alias{check_fun_groupsparsw}
\title{check_fun_groupsparsw}
\usage{
check_fun_groupsparsw(
  X,
  lambda,
  nlambda,
  index,
  sizegroup,
  itermaxw,
  scaling,
  verbose
)
}
\description{
check_fun_groupsparsw
}
\keyword{internal}
