% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsewkm_L0.R
\name{sparsewkm_L0}
\alias{sparsewkm_L0}
\title{L0 Sparse weighted k-means}
\usage{
sparsewkm_L0(
  X,
  centers,
  mode = "s",
  grid = NULL,
  nlambda = 20,
  nstart = 10,
  itermaxw = 20,
  itermaxkm = 10,
  verbose = 1,
  epsilonw = 1e-04
)
}
\arguments{
\item{X}{a dataframe of dimension \code{n} (observations) by \code{p} (variables) with
numerical data.}

\item{centers}{an integer representing the number of clusters.}

\item{mode}{either "s" or "lambda".}

\item{grid}{a vector of numerical values (or a single value) providing 
a grid of values for the regularization parameter.
If NULL (by default), a grid of integer values between 1 and \code{maxs} when \code{mode=s},
and a grid of  \code{nlambda} values in [0,\code{maxlambda}] when  \code{mode=lambda}.}

\item{nlambda}{The size of the grid of values for the regularization parameter (only
when  \code{mode=lambda}).}

\item{nstart}{an integer representing the number of random starts in the k-means algorithm.
By default, \code{nstart=10}.}

\item{itermaxw}{an integer indicating the maximum number of iterations for the inside 
loop over the weights \code{w}. By default, \code{itermaxw=20}.}

\item{itermaxkm}{an integer representing the maximum number of iterations in the k-means 
algorithm. By default, \code{itermaxkm=10}.}

\item{verbose}{an integer value. If \code{verbose=0}, the function stays silent, if \code{verbose=1} (default option), it  prints
whether the stopping criterion over the weights \code{w} is satisfied.}

\item{epsilonw}{a positive numerical value. It provides the precision of the stopping 
criterion over \code{w}. By default, \code{epsilonw =1e-04}.}
}
\value{
A list with the following elements:
 \describe{
 \item{grid}{a numerical vector containing the regularization parameters 
 (a grid of values).}
 \item{W}{a \code{p} by \code{length(grid)} matrix. It contains the weights (0 or 1) 
 associated to each variable.}
\item{cluster}{a \code{n} by \code{length(grid)} integer matrix. It contains the 
cluster memberships, for each value of the regularization parameter.}
\item{bss.per.feature}{a matrix of size \code{p} by \code{length(grid)}. 
It contains the between-class variance computed on the \code{p} variables.}
\item{Z}{a matrix of size \code{n} by \code{p} containing the scaled data 
if \code{scaling=TRUE}, and a copy of \code{X} otherwise.}
}
}
\description{
Sparse weighted k-means with L0 penalty. Two modes are proposed :
one with the regularization parameter called \code{s} (proposed in Chang & al. 2018),
and one with the regularization parameter called \code{lambda}. Here, the data are always 
standardized before clustering. Otherwise, variables with high variance are naturally selected.
}
\examples{
# Swiss data:
data("swiss")
out <- sparsewkm_L0(X = swiss, centers = 3, mode = "s")
# grid of regularization parameters
out$s
# weights of the variables 
out$W

plot(out, what = "weights.features")
plot(out, what = "expl.var")

# partitioning obtained with s=2 selected variables
out$cluster[ , 2]
# between-class variance on each variable
out$bss.per.feature[, 2]
# between-class variance
sum(out$bss.per.feature[, 2])

}
